Added Database resource type support for PList
setResource method now accepts Database resources (simply Database
object from Kohana's Database abstraction layer) so you can use
return value of following method in setResource.

	public function getApps()
	{
	    $this->db->select('name, default_group, create_date');
	    $this->db->from($this->tableNameApps);

	    return $this->db;
	}

note that there is no ->get() to fetch the result.
following code shows how to set the resource

	$applications->setResource($this->Applications->getApps());
	$applications->setLimit(Arag_Config::get('limit', 0));
	$applications->addColumn('name', _("Name"));
        $applications->addColumn('default_group', _("Default Group"));

straightforward, yeah?

be aware that setResource accepts second argument in cases that the
resource is not a simple query and contains join and complex where
conditions. In such case you have to create another resource to
help PList to calculate total number of rows.

In above example, PList executes following query to get total number of
rows.

	$this->db->select('count(*) as count')->from($this->tableNameApps)
             ->get()->current()->count;

which will not work correctly if the main query is a complex query, so be careful.